# perf(Rule): RHICOMPL-3455 make slug generation manual upon import