First I scoped the index on the `slug` column in the `rules` table to the `benchmark_id`, similarly to the `ref_id` from which the `slug` is generated. Then disabled the automatic generation of slugs and explicitly set it in the import and in the factory for testing. 

For verification I ran these two lines after the migration:
```ruby
Rule.all.all? { |r| r.ref_id.parameterize == r.slug }
=> true

Rule.all.all? { |r| r.ref_id.gsub('.', '-').downcase == r.slug }
=> true
```

I tested it by calling the `time bin/rake db:truncate_all ssg:import_rhel_supported` to ensure that the data is being imported into a clean DB. Additionally I had a benchmark measurement to determine the increase of importing rules per benchmark:
```
diff --git a/app/services/concerns/xccdf/util.rb b/app/services/concerns/xccdf/util.rb
index b35a5b3..a6db9d6 100644
--- a/app/services/concerns/xccdf/util.rb
+++ b/app/services/concerns/xccdf/util.rb
@@ -28,7 +28,7 @@ def save_all_benchmark_info
         save_benchmark
         save_profiles
         save_rule_groups
-        save_rules
+        puts ::Benchmark.measure { save_rules }
         save_rule_identifiers
         save_rule_group_rules
         save_rule_group_relationships
```

**Before:**
```
  3.288007   0.121331   3.409338 (  3.952314)
  3.015316   0.077896   3.093212 (  3.557830)
  2.335994   0.088852   2.424846 (  2.821602)
  2.659920   0.089415   2.749335 (  3.216870)
  2.508471   0.077821   2.586292 (  3.095009)
  2.462159   0.073040   2.535199 (  3.032042)
  3.100927   0.076332   3.177259 (  3.668462)
  2.410110   0.066642   2.476752 (  2.884569)
  2.461235   0.085722   2.546957 (  2.997619)
  2.393435   0.062516   2.455951 (  2.881214)
  3.248102   0.072770   3.320872 (  3.813868)
  3.219182   0.077035   3.296217 (  3.754091)
  2.268072   0.057990   2.326062 (  2.725745)
  3.035694   0.092401   3.128095 (  3.584875)
  2.201708   0.056239   2.257947 (  2.649356)
  2.304848   0.057223   2.362071 (  2.768846)
  2.214024   0.062423   2.276447 (  2.690815)
  1.759331   0.051427   1.810758 (  2.153006)
  1.840667   0.056146   1.896813 (  2.208603)
  0.935872   0.034562   0.970434 (  1.151124)
  0.819603   0.012495   0.832098 (  1.006485)
  0.888347   0.027742   0.916089 (  1.073581)
  1.036260   0.025768   1.062028 (  1.243275)
  0.924377   0.018552   0.942929 (  1.095550)
  0.964481   0.019955   0.984436 (  1.121714)

real  4m1.967s
user  3m25.835s
sys 0m4.037s
```

**After:**
```
  2.251216   0.059409   2.310625 (  2.540939)
  1.859929   0.043205   1.903134 (  2.154361)
  1.489057   0.042640   1.531697 (  1.735787)
  1.727620   0.034673   1.762293 (  2.012968)
  1.573173   0.044802   1.617975 (  1.890311)
  1.607754   0.035923   1.643677 (  1.895404)
  2.168034   0.042878   2.210912 (  2.483476)
  1.541941   0.045623   1.587564 (  1.852533)
  1.549115   0.035121   1.584236 (  1.848780)
  1.471936   0.033327   1.505263 (  1.761423)
  2.007558   0.049562   2.057120 (  2.295835)
  1.711028   0.038602   1.749630 (  1.972902)
  1.412123   0.034502   1.446625 (  1.670396)
  1.411195   0.033809   1.445004 (  1.672683)
  1.441967   0.030468   1.472435 (  1.707742)
  1.328558   0.032575   1.361133 (  1.586774)
  1.303524   0.022664   1.326188 (  1.607419)
  1.838662   0.033610   1.872272 (  2.091125)
  1.292697   0.017907   1.310604 (  1.509984)
  0.626073   0.006810   0.632883 (  0.730200)
  0.537427   0.006784   0.544211 (  0.651222)
  0.473544   0.010816   0.484360 (  0.564356)
  0.523727   0.006219   0.529946 (  0.615656)
  0.522258   0.012883   0.535141 (  0.624123)
  0.541166   0.011907   0.553073 (  0.628734)

real  3m31.734s
user  3m1.056s
sys 0m3.478s
```

Each import is ~1.2s faster, which wins us additional 30 seconds overall :tada: :zap: 

## Secure Coding Practices Checklist GitHub Link
- https://github.com/RedHatInsights/secure-coding-checklist

## Secure Coding Checklist
- [ ] Input Validation
- [ ] Output Encoding
- [ ] Authentication and Password Management
- [ ] Session Management
- [ ] Access Control
- [ ] Cryptographic Practices
- [ ] Error Handling and Logging
- [ ] Data Protection
- [ ] Communication Security
- [ ] System Configuration
- [ ] Database Security
- [ ] File Management
- [ ] Memory Management
- [x] General Coding Practices